Remove constant refact
