--- runtime/common/src/constants.rs
@@ -1,106 +1,0 @@
-// This file is part of Bifrost.
-
-// Copyright (C) 2019-2021 Liebi Technologies (UK) Ltd.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-//! A set of constant values used for all runtimes in common.
-
-pub mod currency {
-	use frame_support::weights::{
-		constants::{ExtrinsicBaseWeight, WEIGHT_PER_SECOND},
-		WeightToFeeCoefficient, WeightToFeeCoefficients, WeightToFeePolynomial,
-	};
-	use node_primitives::Balance;
-	use smallvec::smallvec;
-	pub use sp_runtime::Perbill;
-
-	pub const BNCS: Balance = 1_000_000_000_000;
-	pub const DOLLARS: Balance = BNCS;
-	pub const CENTS: Balance = DOLLARS / 100; // assume this is worth about a cent.
-	pub const RELAY_CENTS: Balance = DOLLARS / 30_000;
-	pub const MILLICENTS: Balance = CENTS / 1_000;
-	pub const MILLIBNC: Balance = 1_000_000_000;
-	pub const MICROBNC: Balance = 1_000_000;
-	pub const XCM_WEIGHT: u64 = 1_000_000_000;
-
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		items as Balance * 15 * CENTS + (bytes as Balance) * 6 * CENTS
-	}
-
-	pub struct WeightToFee;
-	impl WeightToFeePolynomial for WeightToFee {
-		type Balance = Balance;
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			let p = super::currency::RELAY_CENTS;
-			let q = 10 * Balance::from(ExtrinsicBaseWeight::get());
-			smallvec![WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			}]
-		}
-	}
-
-	fn base_tx_fee() -> Balance {
-		CENTS / 10
-	}
-
-	pub fn ksm_per_second() -> u128 {
-		let base_weight = Balance::from(ExtrinsicBaseWeight::get());
-		let base_tx_per_second = (WEIGHT_PER_SECOND as u128) / base_weight;
-		let fee_per_second = base_tx_per_second * base_tx_fee();
-		fee_per_second / 100
-	}
-}
-
-/// Time.
-pub mod time {
-	use node_primitives::{BlockNumber, Moment};
-	pub const MILLISECS_PER_BLOCK: Moment = 12000;
-	pub const SECS_PER_BLOCK: Moment = MILLISECS_PER_BLOCK / 1000;
-
-	pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
-
-	// 1 in 4 blocks (on average, not counting collisions) will be primary BABE blocks.
-	pub const PRIMARY_PROBABILITY: (u64, u64) = (1, 4);
-
-	pub const EPOCH_DURATION_IN_BLOCKS: BlockNumber = 6 * HOURS;
-	pub const EPOCH_DURATION_IN_SLOTS: u64 = {
-		const SLOT_FILL_RATE: f64 = MILLISECS_PER_BLOCK as f64 / SLOT_DURATION as f64;
-
-		(EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u64
-	};
-
-	// These time units are defined in number of blocks.
-	pub const MINUTES: BlockNumber = 60 / (SECS_PER_BLOCK as BlockNumber);
-	pub const HOURS: BlockNumber = MINUTES * 60;
-	pub const DAYS: BlockNumber = HOURS * 24;
-	pub const WEEKS: BlockNumber = DAYS * 7;
-
-	// The `LeasePeriod` defination from `polkadot`.
-	pub const KUSAMA_LEASE_PERIOD: BlockNumber = 6 * WEEKS;
-	pub const ROCOCO_LEASE_PERIOD: BlockNumber = 1 * DAYS;
-	pub const WESTEND_LEASE_PERIOD: BlockNumber = 28 * DAYS;
-}
-
-pub mod parachains {
-	pub mod karura {
-		pub const ID: u32 = 2000;
-		pub const KAR_KEY: &[u8] = &[0, 128];
-		pub const KUSD_KEY: &[u8] = &[0, 129];
-	}
-}
